added authentication, changes in controllers, changes in web security

+++ server/src/main/java/coursework/server/controllers/BooksController.java
@@ -3,6 +3,7 @@ package coursework.server.controllers;
 import coursework.server.exceptions.BadRequestException;
 import coursework.server.exceptions.NotFoundException;
 import coursework.server.models.Book;
+import coursework.server.models.User;
 import coursework.server.repositories.BooksRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,17 +15,17 @@ import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/books")
+@RequestMapping("/api/books")
 public class BooksController { //to-do insertion, delete, sorting, search book by author e.t.c, add new loan by id
     @Autowired
     BooksRepository booksRepository;
-    @GetMapping()
+    @GetMapping(value = "/get")
     public ResponseEntity<List<Book>> getAllBooks() {
         return new ResponseEntity<>(booksRepository.findAll(), HttpStatus.OK);
     }
 
-    @GetMapping("{id}")
-    public ResponseEntity<Book> getBookById(@PathVariable long id) {
+    @GetMapping(value ="/get/{id}")
+    public ResponseEntity<Book> getBookById(@PathVariable("id") long id) {
         Optional<Book> book = booksRepository.findById(id);
         if (book.isPresent()) {
             return new ResponseEntity<>(book.get(), HttpStatus.OK);
@@ -33,7 +34,7 @@ public class BooksController { //to-do insertion, delete, sorting, search book b
         }
     }
 
-    @PostMapping(
+    @PostMapping(value ="/post",
             consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Book> postBook(@RequestBody Book newBook) {
@@ -45,17 +46,27 @@ public class BooksController { //to-do insertion, delete, sorting, search book b
         }
     }
 
-    @PutMapping(value = "{id}",
+    @PutMapping(value = "/put/{id}",
             consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Book> updateBook(@PathVariable long id, @RequestBody Book updBook) {
-        updBook.setId(id);
-        booksRepository.save(updBook);
-        return new ResponseEntity<>(updBook, HttpStatus.ACCEPTED);
+    public ResponseEntity<Book> updateBook(@PathVariable("id") long id, @RequestBody Book updBook) {
+        Optional<Book> curBook = booksRepository.findById(id);
+        if (curBook.isPresent()) {
+            updBook.setId(id);
+            booksRepository.save(updBook);
+            return new ResponseEntity<>(updBook, HttpStatus.ACCEPTED);
+        } else {
+            throw new NotFoundException();
+        }
     }
 
-    @DeleteMapping("{id}")
-    public void deleteBookById(@PathVariable long id) {
-        booksRepository.deleteById(id);
+    @DeleteMapping("/delete/{id}")
+    public void deleteBookById(@PathVariable("id") long id) {
+        Optional<Book> curBook = booksRepository.findById(id);
+        if (curBook.isPresent()) {
+            booksRepository.deleteById(id);
+        } else {
+            throw new NotFoundException();
+        }
     }
 }
